requirements.txt
exchangelib>=4.7.1
python-dotenv>=0.19.2
ics>=0.7.2
pytz>=2022.1
cryptography>=37.0.2
